Apt reprocessing and other.

+++ apt.sh
@@ -1,9 +1,19 @@
 
 file=audio/$1
 
-rm -rf $file.png
+if [ `wc -c <${file}.wav` -le 1000000 ]; then
+    echo "Audio file ${file}.wav too small, probably wrong recording"
+    exit
+fi
+
 rm -rf $file_res.wav
+
 sox $file.wav -r 11025 ${file}_res.wav
-/home/franchini/Satellite/wxtoimg/wxtoimg -e HVC ${file}_res.wav $file.png
-convert -rotate 180 $file.png $file.png
+
+/home/franchini/Satellite/wxtoimg/wxtoimg -e ZA -o ${file}_res.wav $file-ZA.png
+convert -rotate 180 $file-ZA.png $file-ZA.png
+
+/home/franchini/Satellite/wxtoimg/wxtoimg -e HVC -o ${file}_res.wav $file-HVC.png
+convert -rotate 180 $file-HVC.png $file-HVC.png
+
 #eog $file.png
